+++ PaxstoreOpenApi/Paxstore/OpenApi/Model/TerminalVariable/Dictionary.cs
@@ -1,6 +1,0 @@
-﻿namespace Paxstore.OpenApi.Model
-{
-    public class Dictionary<T>
-    {
-    }
-}