+++ PaxstoreOpenApi/Paxstore/OpenApi/TerminalVariableApi.cs
@@ -64,20 +64,6 @@ namespace Paxstore.OpenApi
             return result;
         }
 
-      //  private void encryptPasswordVariable(ParameterVariable parameterVariable)
-       // {
-        //    if (string.Equals("P", parameterVariable.Type) && !string.IsNullOrWhiteSpace(parameterVariable.Value))
-          //  {
-            //    try
-              //  {
-                //    parameterVariable.Value = CryptoUtils.byte2hex(CryptoUtils.aesEncrypt(parameterVariable.getValue().getBytes(StandardCharsets.UTF_8), CryptoUtils.encryptMD5(getApiSecret()))));
-    //            }
-      //          catch (Exception ignore)
-        //        {
-          //      }
-          //  }
-        //}
-
         public Result<string> CreateTerminalVariable(TerminalParameterVariableCreateRequest terminalParameterVariableCreateRequest)
         {
             IList<string> validationErrs = new List<string>();
@@ -86,7 +72,7 @@ namespace Paxstore.OpenApi
                 validationErrs.Add(GetMsgByKey("parameterTerminalParameterVariableCreateRequestMandatory"));
                 return new Result<string>(validationErrs);
             }
-            if (string.IsNullOrEmpty(terminalParameterVariableCreateRequest.TID) || string.IsNullOrEmpty(terminalParameterVariableCreateRequest.SerialNo)) {
+            if (string.IsNullOrEmpty(terminalParameterVariableCreateRequest.TID) && string.IsNullOrEmpty(terminalParameterVariableCreateRequest.SerialNo)) {
                 validationErrs.Add(GetMsgByKey("tidAndSnIsMandatory"));
             }
             if (terminalParameterVariableCreateRequest.VariableList == null || terminalParameterVariableCreateRequest.VariableList.Count == 0) {